PP-4884: Use functional style validation
Split each validation into its own function and perform the validation on the
metadata json by composing these functions. This also gives us the possibility
of reusing these functions somewhere else if needed. These functions basically
take a JsonNode, and a list of errors, and appends any errors to the list of
errors if there are any.