Appease IntelliJ's linter in RestClientLoggingFilter
Whilst legal, IntelliJ's linter wasn't a fan of how logstash
`StructuredArgument`s were being used.

This is however fixable by using the `StructuredArgument` parameters as the
values to be interpolated into the logs and `Marker`s for things not
interpolated into the log message.

Replace `String.format` with log placeholders, which should be more performant
and also fixes a bug where the structured log arguments were not being passed
to the logger when logging the start of requests due to a misplaced paren.