PP-10173 Add ApiResponseInstantWithMicrosecondPrecisionSerializer
Add ApiResponseInstantWithMicrosecondPrecisionSerializer, which
is like ApiResponseInstantSerializer except that it uses
microsecond precision rather than millisecond precision. It’s
basically an equivalent of MicrosecondPrecisionDateTimeSerializer
but for Instants rather than ZonedDateTimes.

Also add ISO_LOCAL_DATE_IN_UTC to ApiResponseDateTimeFormatter,
which is a direct copy of the same formatter in connector (and
will eventually replace it).

Remove some @JsonNaming annotations from the serialisers because
they’re not doing anything and we have equivalent annotations on
the JSON classes themselves.